master/requirements.txt
flask==2.0.3
requests==2.26.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
Werkzeug==2.0.3